[2025-11-07#005] Enhance folder, file routers

+++ src/mycampus/routers/files_router.py
@@ -1,7 +1,8 @@
-from fastapi import APIRouter, UploadFile, File
+from fastapi import APIRouter, UploadFile, File, Form
 #APIRouter : FastAPI에서 라우터 객체를 만들고 특정 기능을 묶는 클래스
 #UploadFile : 업로드된 파일을 나타내는 클래스
 #File : 파일 업로드를 처리하기 위한 클래스
+#Form : HTML form 데이터를 처리하기 위한 클래스
 import os
 # os : 운영체제와 상호작용하기 위한 모듈
 
@@ -17,10 +18,25 @@ os.makedirs(base, exist_ok = True)
 #os.makedirs : 지정된 경로에 디렉토리를 생성하는 함수
 #exist_ok=True : 이미 디렉토리가 존재해도 에러를 발생시키지 않음
 
-def upload_file(file: UploadFile = File(...)):
-    save_path = os.path.join(base, file.filename)
+#파일 업로드 함수
+def upload_file(file: UploadFile = File(...),
+                course : str = Form(...),
+                week : str = Form(...),
+                learning_tool : str = Form(...)):
+    #file: UploadFile = File(...) : FastAPI에서 파일 업로드를 처리하기 위한 구문#
+    #Course : str = Form(...) : HTML form 데이터에서 Course 값을 받음
+    #Week : str = Form(...) : HTML form 데이터에서 Week 값을 받음
+    #Learning_tool : str = Form(...) : HTML form 데이터에서 Learning_tool 값을 받음
+
+    folder_path = os.path.join(base, course, week, learning_tool)
     #os.path.join : 여러 경로를 하나의 경로로 조합하는 함수
-    #UploadFile = File(...) : FastAPI에서 파일 업로드를 처리하기 위한 구문
+
+    if not os.path.exists(folder_path):
+        return {"message" : f"Folder '{folder_path}' does not exist."}
+    #os.path.exists : 지정된 경로가 존재하는지 확인하는 함수, 폴더가 없으면 저장하지 않음
+
+    save_path = os.path.join(folder_path, file.filename)
+    #os.path.join : 여러 경로를 하나의 경로로 조합하는 함수, 파일이 실제로 저장되는 전체 경로 생성
 
     with open(save_path,"wb") as uploaded_file:
         uploaded_file.write(file.file.read())
@@ -31,12 +47,8 @@ def upload_file(file: UploadFile = File(...)):
     #파일 저장 성공 메시지 반환
 
 router.add_api_route(
-    path = "/upload",
-    #요청 주소
-    endpoint = upload_file,
-    #실행할 함수
-    methods = ["POST"],
-    #HTTP 메소드
-    summary = "upload file"
-    #swagger ui에 표시될 이름
+    path = "/upload", #요청 주소
+    endpoint = upload_file, #실행할 함수
+    methods = ["POST"], #HTTP 메소드
+    summary = "upload file" #swagger ui에 표시될 이름
 )
